requirements.txt
django==4.2
djangorestframework
Pillow
django-cors-headers
drf-yasg
